+++ src/sqlglot_maxcompute/__init__.py
@@ -1,4 +1,4 @@
-from sqlglot_maxcompute.maxcompute import MaxCompute
+from sqlglot_maxcompute.dialect import MaxCompute
 from sqlglot_maxcompute.parser import MaxComputeParser
 from sqlglot_maxcompute.generator import MaxComputeGenerator